Chapter-18/mlops-images-app/requirements.txt
google-cloud-aiplatform
tensorflow>=2.0.0
numpy
argparse
google-cloud-storage
